Eliminación de líneas comentadas superfluas y corrimiento de varias acciones asíncronas a adminActions
