+++ client/src/redux/helper/axiosBearerToken.js
@@ -1,0 +1,19 @@
+// import axios from 'axios';
+
+// axios.defaults.baseURL = 'http://localhost:3001/'
+// axios.defaults.headers.common =  {'Authorization': `bearer ${token}`}
+
+// export default axios
+
+
+
+// const instance = axios.create({
+//     baseURL: 'https://localhost:3001/api',
+//     timeout: 10000,
+//     headers: {'Authorization': `bearer ${token}`}
+// });
+
+// instance.get('/users')
+// .then(response => {
+//     return response.data
+// })
