+++ client/src/redux/slices/adminActions.js
@@ -3,8 +3,6 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 
 export const getAdmins = createAsyncThunk("admin/getAdmin", async (token) => {
   try {
-    // const response = await axios.get('https://api-paranoid-bikes-production.up.railway.app/api/users')
-    // const response = await axios.get("http://localhost:3001/api/users");
     const response = await axios.get(
       "https://paranoid-bikes-backend.onrender.com/api/admin",
       {
@@ -28,8 +26,6 @@ export const getAdmins = createAsyncThunk("admin/getAdmin", async (token) => {
 
 export const getAdmin = createAsyncThunk("admin/getAdmin", async (id) => {
   try {
-    // const response = await axios.get(`https://api-paranoid-bikes-production.up.railway.app/api/users/${id}`)
-    // const response = await axios.get(`http://localhost:3001/api/users/${id}`);
     const response = await axios.get(
       `${process.env.REACT_APP_URL}/api/admin/${id}`
     );
@@ -45,8 +41,6 @@ export const postAdmin = createAsyncThunk(
     try {
       const response = await axios({
         method: "post",
-        // url: "https://api-paranoid-bikes-production.up.railway.app/api/users",
-        // url: "http://localhost:3001/api/users",
         url: `${process.env.REACT_APP_URL}/api/admin`,
         data: newUser,
       });
@@ -61,9 +55,7 @@ export const putAdmin = createAsyncThunk(
   "admin/putAdmin",
   async ({ _id, ...user }) => {
     try {
-      // const response = await axios.put(`https://api-paranoid-bikes-production.up.railway.app/api/users/${_id}`, pacient)
       const response = await axios.put(
-        // `http://localhost:3001/api/users/${_id}`,
         `${process.env.REACT_APP_URL}/api/admin/${_id}`,
         user
       );
@@ -76,9 +68,7 @@ export const putAdmin = createAsyncThunk(
 
 export const deleteAdmin = createAsyncThunk("admin/deleteAdmin", async (id) => {
   try {
-    // const response = await axios.delete(`https://api-paranoid-bikes-production.up.railway.app/api/users/${id}`)
     const response = await axios.delete(
-      // `http://localhost:3001/api/userks/${id}`
       `${process.env.REACT_APP_URL}/api/admin/${id}`
     );
     return response.data;
@@ -86,6 +76,40 @@ export const deleteAdmin = createAsyncThunk("admin/deleteAdmin", async (id) => {
     return error.message;
   }
 });
+
+export const postProduct = createAsyncThunk(
+  "products/postProduct",
+  async (newBike) => {
+    try {
+      // const config = {
+      //   headers: {Authorization: `Bearer ${token}`}
+      // }
+      const response = await axios.post(
+        `${process.env.REACT_APP_URL}/api/admin/bikes`,
+        newBike
+      );
+      console.log(response.data);
+      return response.data;
+    } catch (error) {
+      return error.message;
+    }
+  }
+);
+
+export const putProduct = createAsyncThunk(
+  "products/putProduct",
+  async ({ id, ...product }) => {
+    try {
+      const response = await axios.put(
+        `${process.env.REACT_APP_URL}/api/bikes/${id}`,
+        product
+      );
+      return response.data;
+    } catch (error) {
+      return error.message;
+    }
+  }
+);
 
 export const deleteProduct = createAsyncThunk(
   "admin/deleteProduct",
@@ -115,11 +139,6 @@ export const updateStock = createAsyncThunk(
       const payload = {
         stock: count,
       };
-      console.log(id.id);
-      console.log(count);
-      console.log(token);
-      console.log(config);
-      console.log(payload);
       const response = await axios.put(
         `${process.env.REACT_APP_URL}/api/admin/bikes/${id.id}`,
         payload,
@@ -131,3 +150,33 @@ export const updateStock = createAsyncThunk(
     }
   }
 );
+
+export const getUsers = createAsyncThunk("users/getUsers", async (token) => {
+  try {
+    const response = await axios.get(`${process.env.REACT_APP_URL}/api/users`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    const data = response.data;
+    // .sort(function(a, b) {
+    //     if(a.first_name < b.first_name) return -1;
+    //     if(a.first_name > b.first_name) return 1;
+    //     return 0
+    // })
+    return data;
+  } catch (error) {
+    return error.message;
+  }
+});
+
+export const deleteUser = createAsyncThunk("users/deleteUser", async (id) => {
+  try {
+    const response = await axios.delete(
+      `${process.env.REACT_APP_URL}/api/users/${id}`
+    );
+    return response.data;
+  } catch (error) {
+    return error.message;
+  }
+});
