+++ client/src/redux/slices/reviewsActions.js
@@ -3,8 +3,6 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 
 export const getReviews = createAsyncThunk("reviews/getReviews", async () => {
   try {
-    // const response = await axios.get('https://api-paranoid-bikes-production.up.railway.app/api/reviews')
-    // const response = await axios.get("http://localhost:3001/api/reviews");
     const response = await axios.get(`${process.env.REACT_APP_URL}/api/reviews`);
     const data = response.data.sort(function (a, b) {
       if (a.author.first_name < b.author.first_name) return -1;
@@ -21,9 +19,7 @@ export const postReview = createAsyncThunk(
   "reviews/postReview",
   async (newReview) => {
     try {
-      // const response = axios.post('https://api-paranoid-bikes-production.up.railway.app/api/reviews', newQuery)
       const response = axios.post(
-        // "http://localhost:3001/api/reviews",
         `${process.env.REACT_APP_URL}/api/reviews`,
         newReview
       );
@@ -38,9 +34,7 @@ export const putReview = createAsyncThunk(
   "reviews/putReview",
   async (id, review) => {
     try {
-      // const response = axios.put(`https://api-paranoid-bikes-production.up.railway.app/api/reviews/${id}`, query)
       const response = axios.put(
-        // `http://localhost:3001/api/reviews/${id}`,
         `${process.env.REACT_APP_URL}api/reviews/${id}`,
         review
       );
@@ -55,7 +49,6 @@ export const deleteReview = createAsyncThunk(
   "reviews/deleteReview",
   async (id) => {
     try {
-      // const response = axios.delete(`https://api-pro-fy-production.up.railway.app/api/reviews/${id}`)
       const response = axios.delete(`${process.env.REACT_APP_URL}/api/reviews/${id}`);
       return response.data;
     } catch (error) {
